+++ sdk/net/basic/tls/dtls_client_socket2.cpp
@@ -14,25 +14,12 @@
 #include <sdk/base/unittest/trace.hpp>
 #include <sdk/crypto/basic/openssl_prng.hpp>
 #include <sdk/net/basic/tls/dtls_client_socket2.hpp>
-#include <sdk/net/tls/tls/extension/tls_extension_ec_point_formats.hpp>
-#include <sdk/net/tls/tls/extension/tls_extension_key_share.hpp>
-#include <sdk/net/tls/tls/extension/tls_extension_psk_key_exchange_modes.hpp>
-#include <sdk/net/tls/tls/extension/tls_extension_renegotiation_info.hpp>
-#include <sdk/net/tls/tls/extension/tls_extension_signature_algorithms.hpp>
-#include <sdk/net/tls/tls/extension/tls_extension_sni.hpp>
-#include <sdk/net/tls/tls/extension/tls_extension_supported_groups.hpp>
-#include <sdk/net/tls/tls/extension/tls_extension_supported_versions.hpp>
-#include <sdk/net/tls/tls/extension/tls_extension_unknown.hpp>
-#include <sdk/net/tls/tls/handshake/tls_handshake_client_hello.hpp>
-#include <sdk/net/tls/tls/handshake/tls_handshake_client_key_exchange.hpp>
-#include <sdk/net/tls/tls/handshake/tls_handshake_finished.hpp>
+#include <sdk/net/basic/tls/tls_composer.hpp>
 #include <sdk/net/tls/tls/record/dtls13_ciphertext.hpp>
 #include <sdk/net/tls/tls/record/tls_record_ack.hpp>
 #include <sdk/net/tls/tls/record/tls_record_alert.hpp>
 #include <sdk/net/tls/tls/record/tls_record_application_data.hpp>
 #include <sdk/net/tls/tls/record/tls_record_builder.hpp>
-#include <sdk/net/tls/tls/record/tls_record_change_cipher_spec.hpp>
-#include <sdk/net/tls/tls/record/tls_record_handshake.hpp>
 #include <sdk/net/tls/tls/tls.hpp>
 #include <sdk/net/tls/tls_advisor.hpp>
 #include <sdk/net/tls/types.hpp>
@@ -40,7 +27,7 @@
 namespace hotplace {
 namespace net {
 
-dtls_client_socket2::dtls_client_socket2(tls_version_t minver) : async_client_socket(), _minver(minver) {
+dtls_client_socket2::dtls_client_socket2(tls_version_t version) : async_client_socket(), _version(version) {
     auto session = &_session;
     session->set_type(session_dtls);
 }
@@ -84,9 +71,9 @@ return_t dtls_client_socket2::do_handshake() {
     return_t ret = errorcode_t::success;
     const size_t bufsize = (1 << 16);
     char buffer[bufsize];
+    tls_advisor* tlsadvisor = tls_advisor::get_instance();
 
     __try2 {
-        tls_advisor* tlsadvisor = tls_advisor::get_instance();
         size_t cbsent = 0;
         binary_t bin;
         uint32 session_status = 0;
@@ -95,151 +82,16 @@ return_t dtls_client_socket2::do_handshake() {
         auto& protection = session->get_tls_protection();
         tls_direction_t dir = from_client;
 
-        // client hello
-        {
-            tls_record_builder builder;
-            auto record = builder.set(session).set(tls_content_type_handshake).set(dir).construct().build();
-            tls_handshake_client_hello* handshake = nullptr;
+        tls_composer composer(session);
+        composer.set_minver(_version);
+        composer.set_maxver(_version);
 
-            handshake = new tls_handshake_client_hello(session);
+        auto lambda = [&](binary_t& bin) -> void {
+            ret = async_client_socket::sendto((char*)&bin[0], bin.size(), &cbsent, (sockaddr*)&_sa, sizeof(_sa));
+            bin.clear();
+        };
 
-            // random
-            {
-                openssl_prng prng;
-
-                binary_t random;  // gmt_unix_time(4 bytes) + random(28 bytes)
-                time_t gmt_unix_time = time(nullptr);
-                binary_append(random, gmt_unix_time, hton64);
-                random.resize(sizeof(uint32));
-                binary_t temp;
-                prng.random(temp, 28);
-                binary_append(random, temp);
-                handshake->set_random(random);
-            }
-
-            bool enable_etm = false;
-            if (session->get_keyvalue().get(session_enable_encrypt_then_mac)) {
-                enable_etm = true;
-            }
-
-            // cipher suites
-            {
-                auto lambda_cs = [&](const tls_cipher_suite_t* cs) -> void {
-                    if (cs->version >= _minver) {
-                        if (enable_etm && (dtls_12 == _minver)) {
-                            if (cbc == cs->mode) {
-                                handshake->add_ciphersuite(cs->code);
-                            }
-                        } else {
-                            handshake->add_ciphersuite(cs->code);
-                        }
-                    }
-                };
-                tlsadvisor->enum_cipher_suites(lambda_cs);
-            }
-            {
-                handshake->get_extensions().add(new tls_extension_renegotiation_info(session));
-                handshake->get_extensions().add(new tls_extension_unknown(tls_ext_session_ticket, session));
-                if (enable_etm && (dtls_12 == _minver)) {
-                    handshake->get_extensions().add(new tls_extension_unknown(tls_ext_encrypt_then_mac, session));
-                }
-                // handshake->get_extensions().add(new tls_extension_unknown(tls_ext_extended_master_secret, session));
-            }
-            {
-                // RFC 9325 4.2.1
-                // Note that [RFC8422] deprecates all but the uncompressed point format.
-                // Therefore, if the client sends an ec_point_formats extension, the ECPointFormatList MUST contain a single element, "uncompressed".
-                auto ec_point_formats = new tls_extension_ec_point_formats(session);
-                (*ec_point_formats).add("uncompressed");
-                handshake->get_extensions().add(ec_point_formats);
-            }
-            {
-                // Clients and servers SHOULD support the NIST P-256 (secp256r1) [RFC8422] and X25519 (x25519) [RFC7748] curves
-                auto supported_groups = new tls_extension_supported_groups(session);
-                (*supported_groups).add("x25519").add("secp256r1").add("x448").add("secp521r1").add("secp384r1");
-                handshake->get_extensions().add(supported_groups);
-            }
-            {
-                auto signature_algorithms = new tls_extension_signature_algorithms(session);
-                (*signature_algorithms)
-                    .add("ecdsa_secp256r1_sha256")
-                    .add("ecdsa_secp384r1_sha384")
-                    .add("ecdsa_secp521r1_sha512")
-                    .add("ed25519")
-                    .add("ed448")
-                    .add("rsa_pkcs1_sha256")
-                    .add("rsa_pkcs1_sha384")
-                    .add("rsa_pkcs1_sha512")
-                    .add("rsa_pss_pss_sha256")
-                    .add("rsa_pss_pss_sha384")
-                    .add("rsa_pss_pss_sha512")
-                    .add("rsa_pss_rsae_sha256")
-                    .add("rsa_pss_rsae_sha384")
-                    .add("rsa_pss_rsae_sha512");
-                handshake->get_extensions().add(signature_algorithms);
-            }
-            {
-                auto supported_versions = new tls_extension_client_supported_versions(session);
-                if (_minver == dtls_13) {
-                    (*supported_versions).add(dtls_13);
-                }
-                (*supported_versions).add(dtls_12);
-                handshake->get_extensions().add(supported_versions);
-            }
-            {
-                auto psk_key_exchange_modes = new tls_extension_psk_key_exchange_modes(session);
-                (*psk_key_exchange_modes).add("psk_dhe_ke");
-                handshake->get_extensions().add(psk_key_exchange_modes);
-            }
-            {
-                auto key_share = new tls_extension_client_key_share(session);
-                (*key_share).add("x25519");
-                handshake->get_extensions().add(key_share);
-            }
-
-            *record << handshake;
-            ret = record->write(dir, bin);
-            record->release();
-        }  // end of client hello
-
-        if (errorcode_t::success != ret) {
-            __leave2;
-        }
-
-        uint32 server_session_status =
-            session_status_server_hello | session_status_server_cert | session_status_server_key_exchange | session_status_server_hello_done;
-        session->wait_change_session_status(server_session_status, get_wto());
-        session_status = session->get_session_status();
-
-        if (0 == (session_status & server_session_status)) {
-            ret = error_handshake;
-            __leave2;
-        }
-
-        // client_key_exchange
-        {
-            tls_record_builder builder;
-            auto record = builder.set(session).set(tls_content_type_handshake).set(dir).construct().build();
-            *record << new tls_handshake_client_key_exchange(session);
-            record->write(dir, bin);
-            record->release();
-        }
-        // change_cipher_spec
-        {
-            tls_record_change_cipher_spec record(session);
-            record.write(dir, bin);
-        }
-        // client finished
-        {
-            tls_record_builder builder;
-            auto record = builder.set(session).set(tls_content_type_handshake).set(dir).construct().build();
-            *record << new tls_handshake_finished(session);
-            record->write(dir, bin);
-            record->release();
-        }
-
-        ret = async_client_socket::sendto((char*)&bin[0], bin.size(), &cbsent, (sockaddr*)&_sa, sizeof(_sa));
-        bin.clear();
+        ret = composer.handshake(from_client, get_wto(), lambda);
     }
     __finally2 {}
 
@@ -353,7 +205,27 @@ return_t dtls_client_socket2::do_secure() {
 
 return_t dtls_client_socket2::do_shutdown() {
     return_t ret = errorcode_t::success;
-    __try2 {}
+    __try2 {
+        auto session = &_session;
+
+        binary_t bin;
+        size_t cbsent = 0;
+
+        {
+            auto dir = from_client;
+            tls_record_builder builder;
+            auto record = builder.set(session).set(tls_content_type_alert).set(dir).construct().build();
+
+            *record << new tls_record_alert(session, tls_alertlevel_warning, tls_alertdesc_close_notify);
+            record->write(dir, bin);
+            record->release();
+        }
+
+        ret = async_client_socket::sendto((char*)&bin[0], bin.size(), &cbsent, (sockaddr*)&_sa, sizeof(_sa));
+
+        session->wait_change_session_status(session_status_server_close_notified, get_wto());
+        auto session_status = session->get_session_status();
+    }
     __finally2 {}
     return ret;
 }