+++ sdk/net/basic/tls/dtls_client_socket2.hpp
@@ -35,7 +35,7 @@ namespace net {
  */
 class dtls_client_socket2 : public async_client_socket {
    public:
-    dtls_client_socket2(tls_version_t minver = dtls_12);
+    dtls_client_socket2(tls_version_t version = dtls_12);
 
     virtual return_t sendto(const char* ptr_data, size_t size_data, size_t* cbsent, const struct sockaddr* addr, socklen_t addrlen);
 
@@ -50,7 +50,7 @@ class dtls_client_socket2 : public async_client_socket {
 
    private:
     tls_session _session;
-    tls_version_t _minver;
+    tls_version_t _version;
 
     critical_section _mlock;
     std::queue<bufferqueue_item_t> _mq;