+++ sdk/net/tls/tls/handshake/tls_handshake_server_hello.cpp
@@ -142,7 +142,7 @@ return_t tls_handshake_server_hello::do_preprocess(tls_direction_t dir) {
         if (0 == (session_status_client_hello & session_status)) {
             ret = errorcode_t::error_handshake;
             session->push_alert(dir, tls_alertlevel_fatal, tls_alertdesc_unexpected_message);
-            __leave2;
+            __leave2_trace(ret);
         }
     }
     __finally2 {