+++ sdk/net/tls/tls/handshake/tls_handshake_server_hello_done.cpp
@@ -50,6 +50,7 @@ return_t tls_handshake_server_hello_done::do_preprocess(tls_direction_t dir) {
             if (0 == (session_status_server_key_exchange & session_status)) {
                 ret = errorcode_t::error_handshake;
                 session->push_alert(dir, tls_alertlevel_fatal, tls_alertdesc_unexpected_message);
+                __leave2_trace(ret);
                 __leave2;
             }
         }