+++ CHANGELOG.md
@@ -4,8 +4,8 @@
   * [study] QUIC (RFC 9001, 9369)
   * [study] TLS (RFC 4346, 5246, 8446, 8448)
 
-* Revision 766
-  * [maybe] DTLS over UDP, client
+* Revision 766 (Tag 1.23)
+  * [tested] DTLS over UDP, client
 
 * Revision 765